backend/backend/requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-extensions==3.1.1
django-rest-auth==0.9.5
django-rest-multiple-models==2.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-writable-nested==0.6.2
Pillow==8.1.0
PyJWT==2.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
